add dev server
