+++ package.json
@@ -20,7 +20,8 @@
   },
   "scripts": {
     "build": "node build.js",
-    "build:dev": "RUNTIME_API_HOST='http://localhost:9000' RUNTIME_ENV=dev node build.js",
+    "build:dev": "RUNTIME_API_HOST='https://zerobase-api-stage.herokuapp.com' RUNTIME_ENV=dev node build.js",
+    "start:dev": "RUNTIME_API_HOST='https://zerobase-api-stage.herokuapp.com' RUNTIME_ENV=dev node devServer.js",
     "serve": "http-server ./public"
   },
   "devDependencies": {
@@ -31,6 +32,8 @@
     "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "^2.5.0"
+    "eslint-plugin-react-hooks": "^2.5.0",
+    "express": "^4.17.1",
+    "node-watch": "^0.6.3"
   }
 }
